# Reduce memory usage when AEAD en/decrypting large messages